a large rewrite of the use mixed.
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rmetrics@6028 5d4a7380-5c0c-0410-b446-d4a5018c0119

+++ pkg/randtoolbox/src/LowDiscrepancy-halton.c
@@ -92,10 +92,10 @@ void INITHALTON(int DIMEN, double *QUASI)
 }
 
 //compute the radical inverse function of integer OFFSET with base primeNumber[DIMEN]
-double HALTONREC(int DIMEN, int OFFSET)
+double HALTONREC(int DIMEN, unsigned long long OFFSET)
 {
     //in Fortran code BASE contains prime numbers
-    int ITER, DIGIT;
+    unsigned long long ITER, DIGIT;
     double HALF;
     double QUASI;
